Move functionCode request related code to AwsLambda
